PARAMETRA cost_biosecurity data

Information about prices of products needed to implement biosecurity measures

Usage

cost_biosecurity

Format

A data frame.

Source

Compiled from various scientific publications; see the "ref" column.

Details

Columns:

biosecurity_measure

Category of measure

country

Country or region where data was collected

filled_by

Data entry person identifier

filled_date

Data entry date

id

Parameter identifier

last_access

URL last access date

lifespan

Expected lifespan of the product if applicable and based on typical usage

parameter

Parameter measured

price

Price of the product available online without VAT

price_by_event

Total cost per event to implement the biosecurity measure without VAT

product_name

Commercial name of the product

product_type

Specific type of product needed to implement the biosecurity measure

qty_by_event

Quantity of the product used per event to implement the biosecurity measure

ref

Reference: source DOI or URL

ref_last_access

Date reference was automatically checked for the last time

ref_short

Short reference: source author and year

ref_status

Reference status in the last aoutomatic check

remarks

Additional notes or comments

transmission_pathway

Transmission pathway targeted by this biosecurity measure

unit_needed

Unit of the product needed per event to implement the biosecurity measure (e.g. gram, liter, unit)

year_cost

Year the cost was included